% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{super_funkcja_wszystko_dziala}
\alias{super_funkcja_wszystko_dziala}
\title{Print  function}
\usage{
super_funkcja_wszystko_dziala(x)
}
\arguments{
\item{x}{an object of \code{list} class}
}
\value{
\code{NULL} (invisible)
}
\description{
Print  function
}
\details{
Costam robi
}
\seealso{
\code{\link[base]{debug}}
}
